@@ -2,20 +2,24 @@ package com.example.ruslanyussupov.popularmovies.adapters;
 
 
 import android.databinding.DataBindingUtil;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.ruslanyussupov.popularmovies.OnMovieClickListener;
+import com.example.ruslanyussupov.popularmovies.App;
 import com.example.ruslanyussupov.popularmovies.R;
+import com.example.ruslanyussupov.popularmovies.Utils;
 import com.example.ruslanyussupov.popularmovies.data.model.Movie;
 import com.example.ruslanyussupov.popularmovies.databinding.ItemMovieBinding;
-import com.example.ruslanyussupov.popularmovies.utils.NetworkUtils;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
+
+import javax.inject.Inject;
+
 
 public class MovieAdapter extends RecyclerView.Adapter<MovieAdapter.ViewHolder> {
 
@@ -23,16 +27,19 @@ public class MovieAdapter extends RecyclerView.Adapter<MovieAdapter.ViewHolder> 
     private List<Movie> mMovies;
     private final OnMovieClickListener mMovieClickListener;
 
+    @Inject
+    Utils utils;
+
     public MovieAdapter(List<Movie> movies, OnMovieClickListener onMovieClickListener) {
         mMovies = movies;
         mMovieClickListener = onMovieClickListener;
+        App.getComponent().inject(this);
     }
 
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
 
-        // Inflate the our custom layout
         ItemMovieBinding binding = DataBindingUtil.inflate(LayoutInflater.from(parent.getContext()),
                 R.layout.item_movie, parent, false);
 
@@ -58,29 +65,39 @@ public class MovieAdapter extends RecyclerView.Adapter<MovieAdapter.ViewHolder> 
 
             mBinding = binding;
 
-            binding.getRoot().setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    int position = getAdapterPosition();
-                    mMovieClickListener.onMovieClick(mMovies.get(position));
-                }
+            binding.getRoot().setOnClickListener(v -> {
+                int position = getAdapterPosition();
+                mMovieClickListener.onMovieClick(mMovies.get(position));
             });
 
         }
 
         void bind(Movie movie) {
             mBinding.executePendingBindings();
-            // Load movie poster into ImageView
-            String posterPath = NetworkUtils.buildMoviePosterUrlPath(movie.getPosterPath());
-            Picasso.get().load(posterPath).placeholder(R.drawable.poster_placeholder)
-                    .error(R.drawable.poster_error)
-                    .into(mBinding.moviePoster);
+            if (utils.hasNetworkConnection()) {
+                Picasso.get().load(movie.getFullPosterPath()).placeholder(R.drawable.poster_placeholder)
+                        .error(R.drawable.poster_error)
+                        .into(mBinding.moviePoster);
+            } else {
+                Bitmap poster = BitmapFactory.decodeFile(movie.getPosterLocalPath());
+
+                if (poster == null) {
+                    mBinding.moviePoster.setImageResource(R.drawable.backdrop_placeholder);
+                } else {
+                    mBinding.moviePoster.setImageBitmap(poster);
+                }
+            }
+
         }
     }
 
     public void updateData(List<Movie> movies) {
         mMovies = movies;
         notifyDataSetChanged();
+    }
+
+    public interface OnMovieClickListener {
+        void onMovieClick(Movie movie);
     }
 
 }
