@@ -1,18 +1,19 @@
-package com.example.ruslanyussupov.popularmovies.ui;
+package com.example.ruslanyussupov.popularmovies;
 
 
 import android.graphics.Rect;
+import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 
-class ItemDecoration extends RecyclerView.ItemDecoration {
+public class ItemDecoration extends RecyclerView.ItemDecoration {
 
     private final int offsetLeft;
     private final int offsetTop;
     private final int offsetRight;
     private final int offsetBottom;
 
-    ItemDecoration(int offsetLeft, int offsetTop, int offsetRight, int offsetBottom) {
+    public ItemDecoration(int offsetLeft, int offsetTop, int offsetRight, int offsetBottom) {
         this.offsetLeft = offsetLeft;
         this.offsetTop = offsetTop;
         this.offsetRight = offsetRight;
@@ -20,7 +21,7 @@ class ItemDecoration extends RecyclerView.ItemDecoration {
     }
 
     @Override
-    public void getItemOffsets(Rect outRect, View view, RecyclerView parent, RecyclerView.State state) {
+    public void getItemOffsets(@NonNull Rect outRect, @NonNull View view, @NonNull RecyclerView parent, @NonNull RecyclerView.State state) {
         super.getItemOffsets(outRect, view, parent, state);
         outRect.set(offsetLeft, offsetTop, offsetRight, offsetBottom);
     }
