@@ -1,36 +1,34 @@
-package com.example.ruslanyussupov.popularmovies.ui;
+package com.example.ruslanyussupov.popularmovies.list;
 
 
-import android.arch.lifecycle.Observer;
 import android.arch.lifecycle.ViewModelProviders;
 import android.content.Context;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.GridLayoutManager;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.ruslanyussupov.popularmovies.OnMovieClickListener;
+import com.example.ruslanyussupov.popularmovies.ItemDecoration;
 import com.example.ruslanyussupov.popularmovies.R;
 import com.example.ruslanyussupov.popularmovies.adapters.MovieAdapter;
-import com.example.ruslanyussupov.popularmovies.data.model.Movie;
 import com.example.ruslanyussupov.popularmovies.databinding.FragmentMovieGridBinding;
-import com.example.ruslanyussupov.popularmovies.utils.NetworkUtils;
 
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Collections;
+
+import timber.log.Timber;
+
+import static com.example.ruslanyussupov.popularmovies.adapters.MovieAdapter.*;
 
 
 public class MovieGridFragment extends Fragment {
 
-    private static final String LOG_TAG = MovieGridFragment.class.getSimpleName();
-
-    // Bundle keys for saving instance state
     public static final String EXTRA_MOVIE = "movie";
 
     private static final int MOVIE_GRID_COLUMNS = 2;
@@ -38,7 +36,7 @@ public class MovieGridFragment extends Fragment {
     private MovieAdapter mMovieAdapter;
     private OnMovieClickListener mMovieClickListener;
     private FragmentMovieGridBinding mBinding;
-    private MovieGridViewModel mViewModel;
+    private MainViewModel mViewModel;
 
     public MovieGridFragment() {}
 
@@ -57,10 +55,8 @@ public class MovieGridFragment extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
 
-        Log.d(LOG_TAG, "onCreateView");
-
+        Timber.d("onCreateView");
         mBinding = DataBindingUtil.inflate(inflater, R.layout.fragment_movie_grid, container, false);
-
         return mBinding.getRoot();
 
     }
@@ -69,60 +65,62 @@ public class MovieGridFragment extends Fragment {
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        Log.d(LOG_TAG, "onActivityCreated");
+        Timber.d("onActivityCreated");
 
-        // Set up MovieAdapter and GridLayoutManager to RV
-        mMovieAdapter = new MovieAdapter(new ArrayList<Movie>(), mMovieClickListener);
+        mMovieAdapter = new MovieAdapter(new ArrayList<>(), mMovieClickListener);
         mBinding.rvMovies.setAdapter(mMovieAdapter);
         mBinding.rvMovies.setLayoutManager(new GridLayoutManager(getActivity(), MOVIE_GRID_COLUMNS));
         int offset = getResources().getDimensionPixelOffset(R.dimen.movie_item_offset);
         mBinding.rvMovies.addItemDecoration(new ItemDecoration(offset, offset, offset, offset));
 
-        mViewModel = ViewModelProviders.of(getActivity()).get(MovieGridViewModel.class);
+        mViewModel = ViewModelProviders.of(this).get(MainViewModel.class);
 
-        if (NetworkUtils.hasNetworkConnection(getActivity())) {
-            mViewModel.getMovies().observe(this, new Observer<List<Movie>>() {
-                @Override
-                public void onChanged(@Nullable List<Movie> movies) {
-                    if (movies == null || movies.isEmpty()) {
-                        showEmptyState();
+        mViewModel.getResultLiveData().observe(this, result -> {
+
+            mBinding.rvMovies.scrollToPosition(0);
+
+            switch (result.state) {
+                case LOADING:
+                    Timber.d("Movies loading...");
+                    mBinding.loadingPb.setVisibility(View.VISIBLE);
+                    break;
+                case SUCCESS:
+                    mBinding.loadingPb.setVisibility(View.GONE);
+                    if (result.data == null || result.data.isEmpty()) {
+                        Timber.d("Result is empty.");
+                        showSnackBar("No movies!");
+                        mMovieAdapter.updateData(Collections.emptyList());
                     } else {
-                        showMovies();
-                        mMovieAdapter.updateData(movies);
+                        Timber.d("Movies loaded successfully: %s", result.data);
+                        mMovieAdapter.updateData(result.data);
                     }
-                }
-            });
-        } else {
-            showNoNetworkConnectionState();
-        }
+                    break;
+                case ERROR:
+                    mBinding.loadingPb.setVisibility(View.GONE);
+                    if (mViewModel.getUtils().hasNetworkConnection()) {
+                        Timber.e("Error while loading movies: %s", result.error);
+                        showSnackBar(result.error);
+                    } else {
+                        Timber.w("No network connection.");
+                        showSnackBar("No connection!");
+                    }
+                    break;
+            }
 
+        });
 
     }
 
     @Override
     public void onDestroy() {
-        Log.d(LOG_TAG, "onDestroy");
+        Timber.d("onDestroy");
         super.onDestroy();
     }
 
-    private void showEmptyState() {
-        mBinding.loadingPb.setVisibility(View.GONE);
-        mBinding.rvMovies.setVisibility(View.GONE);
-        mBinding.stateTv.setVisibility(View.VISIBLE);
-        mBinding.stateTv.setText(R.string.empty_state);
-    }
-
-    private void showMovies() {
-        mBinding.loadingPb.setVisibility(View.GONE);
-        mBinding.stateTv.setVisibility(View.GONE);
-        mBinding.rvMovies.setVisibility(View.VISIBLE);
-    }
-
-    private void showNoNetworkConnectionState() {
-        mBinding.loadingPb.setVisibility(View.GONE);
-        mBinding.rvMovies.setVisibility(View.GONE);
-        mBinding.stateTv.setVisibility(View.VISIBLE);
-        mBinding.stateTv.setText(R.string.no_network_connection_state);
+    private void showSnackBar(String message) {
+        Snackbar.make(mBinding.rvMovies, message, Snackbar.LENGTH_LONG)
+                .setAction("Retry", listener -> mViewModel.retry())
+                .show();
     }
 
 }
